+++ packages/camera/camera_android_camerax/lib/src/camerax_library.g.dart
@@ -123,6 +123,7 @@ class PigeonOverrides {
     ResolutionSelector? resolutionSelector,
     int? targetRotation,
     CameraXFlashMode? flashMode,
+    int? jpegQuality,
   })?
   imageCapture_new;
 
@@ -5118,12 +5119,14 @@ class ImageCapture extends UseCase {
     ResolutionSelector? resolutionSelector,
     int? targetRotation,
     CameraXFlashMode? flashMode,
+    int? jpegQuality,
   }) {
     if (PigeonOverrides.imageCapture_new != null) {
       return PigeonOverrides.imageCapture_new!(
         resolutionSelector: resolutionSelector,
         targetRotation: targetRotation,
         flashMode: flashMode,
+        jpegQuality: jpegQuality,
       );
     }
     return ImageCapture.pigeon_new(
@@ -5132,6 +5135,7 @@ class ImageCapture extends UseCase {
       resolutionSelector: resolutionSelector,
       targetRotation: targetRotation,
       flashMode: flashMode,
+      jpegQuality: jpegQuality,
     );
   }
 
@@ -5142,6 +5146,7 @@ class ImageCapture extends UseCase {
     this.resolutionSelector,
     int? targetRotation,
     CameraXFlashMode? flashMode,
+    int? jpegQuality,
   }) : super.pigeon_detached() {
     final int pigeonVar_instanceIdentifier = pigeon_instanceManager
         .addDartCreatedInstance(this);
@@ -5161,6 +5166,7 @@ class ImageCapture extends UseCase {
           resolutionSelector,
           targetRotation,
           flashMode,
+          jpegQuality,
         ]);
     () async {
       final pigeonVar_replyList = await pigeonVar_sendFuture as List<Object?>?;